pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "browser-use>=0.10.1",
    "langchain-openai>=0.2.0",
    "loguru>=0.7.0",
    "pydantic-ai>=1.36.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.45",
    "sqlitedict>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
